requirements.txt
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
blessed==1.20.0
blinker==1.7.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clip
colorama==0.4.6
comm==0.2.1
ConfigArgParse==1.7
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dash==2.16.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
ffmpy==0.3.2
filelock==3.13.1
fire==0.5.0
Flask==3.0.2
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.3.1
ftfy==6.1.3
gpustat==1.1.1
GPUtil==1.4.0
gradio==4.19.0
gradio_client==0.10.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanize==4.9.0
idna==3.6
imageio==2.34.0
importlib_metadata==7.0.2
importlib_resources==6.3.1
ipython
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
lazy_loader==0.3
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4

open3d==0.18.0
orjson==3.9.15
packaging==24.0
pandas==2.2.1
paramiko==3.4.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
plotly==5.19.0
-e git+https://github.com/Voxai2023/point-e.git@a728fcf3eb828792eedadfe3dd91b792287dc443#egg=point_e
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-vox-io==0.1
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
PyNaCl==1.5.0
pyparsing==3.1.2
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.3
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.3.0
tifffile==2024.2.12
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torchvision==0.17.1
tqdm==4.66.2
traitlets==5.14.1
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.28.1
watchfiles==0.21.0
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.0.1
widgetsnbextension==4.0.10
yarl==1.9.4
zipp==3.17.0
